Addd the dataTape file outline and login to calculate the TTY dimensions of a POSIX shell

+++ src/dataTape.c
@@ -4,7 +4,7 @@ void dt_DataTape(DataTape *data)
 {
   data->mDataIndex = 0;
   data->mDataLen = 8;
-  data->mData = calloc((size_t) data->mDataLen, (size_t) sizeof(char));
+  data->mData = calloc(data->mDataLen, sizeof(char));
 }
 
 void dt_DelDataTape(DataTape *data)
